commit by qjc

+++ mmdet/models/detectors/base.py
@@ -167,7 +167,6 @@ class BaseDetector(BaseModule, metaclass=ABCMeta):
         if torch.onnx.is_in_onnx_export():
             assert len(img_metas) == 1
             return self.onnx_export(img[0], img_metas[0])
-
         if return_loss:
             return self.forward_train(img, img_metas, **kwargs)
         else:
@@ -190,7 +189,7 @@ class BaseDetector(BaseModule, metaclass=ABCMeta):
             if isinstance(loss_value, torch.Tensor):
                 log_vars[loss_name] = loss_value.mean()
             elif isinstance(loss_value, list):
-                log_vars[loss_name] = sum(_loss.mean() for _loss in loss_value)
+                log_vars[loss_name] = sum(_loss.float().mean() for _loss in loss_value)
             else:
                 raise TypeError(
                     f'{loss_name} is not a tensor or list of tensors')
